first attempt at damage rules, fixed error in ecs.

+++ Source/Core/Game_Objects.h
@@ -19,7 +19,7 @@ void Object_Create(
     const char* file_path
 );
 
-void* Object_Lookup_Component(u32 object_id, u8 component);
+void* Object_Lookup_Component(i32 object_id, i32 component);
 
 void Object_Add_VAO(GLint data);
 void Object_Add_VBO(GLint data);
